Hi all,

@sstaehler @martinvandriel @tnissen @alexhutko @CTrabant - pinging you all as its somehow relevant to all of us.

I spent some time last week and today trying to make Instaseis (and by extension) Syngine faster. So far I only looked at I/O performance as that is the dominant part and everything else can be made faster by just throwing more CPUs at it.

To extract a single 3 component seismogram we have to access data from 5 arrays across two netCDF files. At each array we have to get data for all 25 GLL points. Worst case this results in 125 read operations for each 3C seismogram.

Currently this data is stored in chunked netCDF files ([more about chunks](https://www.hdfgroup.org/HDF5/doc/Advanced/Chunking/)) at one chunk per GLL point time series (that's what makes it fairly performant even though data is accessed with a loop over the slow index). Nonetheless 125 chunks have to be found (HDF traverses a btree for that) and read.

This PR does a couple of things in an attempt to speed things up.

1. Replace Python `netcdf4` with `h5py`, a Python package to read HDF5 files. Newer netCDF versions are based on HDF5 thus we essentially skip one layer of software which is always nice.
2. Batch I/O operations when using the current database structure. This is a bit faster likely due to less function call overheads and maybe some internal h5py/hdf5 optimizations.
3. Try different HDF5/netCDF structures/settings. I implemented a couple but only intend to keep one - otherwise the maintenance overhead is too much.

The most promising database restructurings are:

* Transpose the big arrays (so the fast index is last - hdf5 uses C memory layout) and store it continuous and uncompressed in the files. Thus adjacent GLL time series can be read with a single read operation. The downside is that the files are a bit bigger (no compression) and things like check-sums don't work anymore.
* Take all 5 arrays from the two files and restructure the file so the data for all GLL points from a single element can be acquired with just a single read operation. Thus data is stored in a single large 5D array (element id, time_axis, gll_id_i, gll_id_j, direction). This means that duplicate GLL points are no longer discarded - the file is also no longer chunked and thus not compressed. This results in much bigger files but also quite a bit faster.

Two things that might still be worthwhile to try are (I'll be pretty busy the next two weeks - please feel free to try it):

* Write the 3 arrays from the horizontal and the 2 from the vertical DBs to single arrays in each file and adjust the chunking. This should already be quite a speedup without increasing the file size at all.
* Use the 5D single file struture and adjust the chunking so a chunk is all the data from one element. This would enable compression and thus smaller files.


What trade-off between file size and extraction speed should we make? Also for the syngine use case. Is doubling the file size ok? Instaseis will always support the current database structure but I will also add support for one more layout. Which one is best?


The following measurements have to be taken with a large grain of salt. It is really hard to make these measurements as multiple layers of caches are active at any point in time (OS caches/HDF5 caches/Instaseis caches/...). If you try this, please clear the operating system disc buffers before making the measurements. Also use a database with a size of at least a couple times your system's memory - the buffers are surprisingly effective and small databases tint the result as we essentially only care about performance for the large databases.

Furthermore they are highly dependent on the system and I/O architecture and load.

I made the measurements with

```bash
python -m instaseis.benchmark --pattern UnbufferedFullyRandom --time=20 database_name
```

This branch also contains a tool to convert databases from the current layout to another layout:

```bash
python /../instaseis/scripts/repack_instaseis_database.py --method=unrolled_merge DB DB_unrolled
```

This can take a very long time!

## Benchmarks
### 2s database over a pretty slow network NAS


###### Current DB Structure (1.1 TB) and current Instaseis Master:
        1.3531 sec/seismogram
        0.739046 seismograms/sec
        
        
###### Current DB Strucuture (1.1 TB) and h5py Branch:
        1.12009 sec/seismogram
        0.892785 seismograms/sec
        
###### Transposed and non-chunked DB (1.4 TB) and h5py Branch:
        0.355843 sec/seismogram
        2.81023 seismograms/sec
        
###### Unrolled and unchunked DB (2.2 TB) and h5py Branch:
        0.239713 sec/seismogram
        4.17166 seismograms/sec

        
        
### 5s mars database on an external USB hard disc

###### Current DB Structure (37 GB) and current Instaseis Master:
	0.446772 sec/seismogram
	2.23828 seismograms/sec
        
###### Current DB Strucuture (37 GB) and h5py Branch:
	0.331634 sec/seismogram
	3.01537 seismograms/sec
        
###### Transposed and non-chunked DB (54 GB) and h5py Branch:
	0.208579 sec/seismogram
	4.79434 seismograms/sec
        
###### Unrolled and unchunked DB (83 GB) and h5py Branch:
	0.0630697 sec/seismogram
	15.8555 seismograms/sec