@sstaehler 2016-01-19
Maybe one usability hint: The transformation script should probably take the original axisem_output.nc4 files as input, so that we save one level of intermediate files.

Also, the script could be moved to the AxiSEM repository.