@sstaehler 2016-01-19
Okay, another critical issue, I noticed is that the resulting file is not valid NetCDF4 file anymore. 

I could not pinpoint the reason yet, but I assume it is due to missing dimensions in the output file. One of the two big differences between NetCDF4 and HDF5 is that the former uses named and shared dimensions (lat, lon, x, y, z, time, depth, etc), while the latter just uses array shapes.
https://www.unidata.ucar.edu/blogs/developer/en/entry/dimensions_scales
https://www.unidata.ucar.edu/software/netcdf/docs/interoperability_hdf5.html#reading_with_hdf5

Since the kerner will use NetCDF4 for the foreseeable future and performance-wise it is just as critical, I would like to have valid NetCDF4 files as an output of the script. I see two possibilities

 1. Use the NetCDF4 module for the conversion script (we do not really care about the performance of the conversion script, do we?)
 2. Attach a dimension scale to each output variable. That seems rather tedious.