@martinvandriel 2016-01-19
First thanks of all for this initiative. My feeling is we might want to move in two directions with different applications in mind:

 - Heavy IO applications like the Kerner, where speed is much more important than storage and where DBs are probably only stored temporarily.

 - Low IO long time applications like the Insight mission, where we need to store many DBs and have few access operations. Here, storage is the crucial part and even some additional CPU overhead with respect to the current master would be acceptable, if that allows additional compression.

My suggestion would be to have at maximum two storage patterns, that are optimized for these two directions.

Anyway, both can be done in postprocessing (currently called 'fieldtransform') in python and don't require changes in the Fortran code, so this keeps things relatively simple.