@sstaehler 2016-01-19
@martinvandriel: agreed, probably, we should move the functionality into fieldtransform.py. This would also solve the NetCDF4 problem.

The IRIS guys could then decide which storage pattern they prefer.