@sstaehler 2016-01-19
>> Okay, another critical issue, I noticed is that the resulting file is not valid NetCDF4 file anymore.
>That should be easy to fix. I think I just have to copy the dimension scales of the datasets - they also exist in hdf5. I'll make sure the files stay valid netCDF files when I polish the script.

If it's easy to do in h5py, then I don't care which library is used. But it's also needed for all the mesh variables, which are now just copied in a loop. 
For testing: If 

    ncdump -hs file.nc4 

can read it, it should be fine.