@martinvandriel 2016-01-19
> You could always use a higher gzip compression level or szip compression. Also make larger chunks which should compress better as compression is per chunk.

These things don't really help, because floats look very much like random numbers, hence have high entropy and don't compress well (only the zeroes before p-arrival compress well). So the only promising strategy I see here would be to use e.g. wavelets or predictor/corrector methods and convert to integers to make things compressable.