@sstaehler 2016-02-01
Dear @alexhutko @CTrabant,

since another code would depend on these changes on my side, I'd like to reach a decision soon. What is your opinion on the trade-off between performance and disk space?