@chad-iris 2016-02-01
Well, based solely on Lion's tests I would have chosen a middle solution (Transposed and non-chunked) to get a performance gain without doubling the file size.  Syngine doesn't really fit into either of Martin's target use cases in that we know the IO is heavy but storage size must be considered.

Let us get farther into testing the variations on our storage system and we'll have more information to provide.  I expect to have some benchmarks in the next couple of days.